+++ application/config/routes.php
@@ -55,10 +55,11 @@ $route['translate_uri_dashes'] = FALSE;
 
 $route['parts/([a-zA-Z][0-9])'] = 'part/display/$1';
 $route['parts'] = 'partsFactory';
+
 $route['about'] = 'about/index';
 $route['teammate/([a-zA-Z]+)'] = 'about/member/$1';
 $route['history'] = "HistoryController/index";
-$route['assembly'] = 'assemblycon';
+$route['assembly'] = 'assembly';