+++ application/models/PartsDatabase.php
@@ -1,0 +1,7 @@
+<?php
+class partsDatabase extends MY_Model {
+        public function __construct()
+        {
+                parent::__construct('parts','partID');
+        }
+}