+++ application/views/assembly.php
@@ -1,48 +1,57 @@
-<div class = 'col-mid12' id = 'content'>
 
-<div id ="body">
-	<p> this is the assembly select your options </p>
-
-<div class ="row">
-<table style="border: 1px solid white; width: 50%; height:100px;">	
-	<tr>
-		<th></th>
-		<th>part</th>
-		<th>Part code</th>
-		<th>Ca Code</th>
-		<th>plant built</th>
-	</tr>
-	{partTable}
-	<tr>
-	<td><input type="checkbox" name="pick[]" value="select"></td>
-		<td>{part}</td>
-		<td>{partCode}</td>
-		<td>{caCode}</td>
-		<td>{plant}</td>
-	</tr>
-	{/partTable}
-</table>
-</div>
-
-<button type= "button" style="color:black;">Build</button> <button type= "button" style="color:black;">return to head office</button> 
-</br>
-<div class ="row">
-<table style="border: 1px solid white; width: 50%; height:100px;">	
-	<tr>
-		<th></th>
-		<th>part</th>
-	</tr>
-	{completedBots}
-	<tr>
-		<td><input type="checkbox" name="pick[]" value="select"></td>
-		<td>{part}</td>
-	</tr>
-	{/completedBots}
-</table>
-</div>
-
-
-<button class = "button" style="color:black;">ship</button>
-
-</div>
-
+    <body>
+		<h1>Assembly<h1>
+        <div class="container">
+            <div id="content">
+				<h2>parts</h2>
+				<h3>filter: <a href="/filter/actor/ALL" ><button style="color:black;">All</button></a> <a href="/filter/actor/A" style="color:black;"><button style="color:black;">A</button></a> <a href="/filter/actor/B"><button style="color:black;">B</button></a> <a href="/filter/actor/C" ><button style="color:black;">C</button></a> <a href="/filter/actor/M" ><button style="color:black;">M</button></a> <a href="/filter/actor/R" ><button style="color:black;">R</button></a> <a href="/filter/actor/W" ><button style="color:black;">W</button></a></h3>
+				<form method='POST' action='/assembly/pressed'>
+					<div class = "row">
+						<span class='col-md-6'>
+						{hasParts}
+							<table class = "table">
+								<tr>
+									<th></th>
+									<th>part</th>
+									<th>Part code</th>
+									<th>Ca code</th>
+									<th>plant made</th>
+								</tr>
+								{partTable}
+								<tr>
+									<td><input type="checkbox" name="pick{part}" /></td>
+									<td>{part}</td>
+									<td>{partCode}</td>
+									<td>{caCode}</td>
+									<td>{plantBuiltAt}</td>
+								</tr>
+								{/partTable}
+							</table>
+						</span>
+					</div>
+					<input type= "submit" name = "build" value= "Build" style="color:black;"/> <input type= "submit" name= "return" value= "return to head office" style="color:black;"/>
+				</form>
+				<h2>completed bots</h2>
+				<div class = "row">	
+					{hasBots}				
+					<table class = "table">
+						<tr>
+							<th></th>
+							<th>model</th>
+							<th>head Ca code</th>
+							<th>torso Ca code</th>
+							<th>bottom Ca made</th>
+						</tr>
+						{completeBots}
+						<tr>
+							<td></td>
+							<td>{model}</td>
+							<td>{headCaCode}</td>
+							<td>{torsoCaCode}</td>
+							<td>{bottomCaCode}</td>
+						</tr>
+						{/completeBots}
+					</table>
+				</div>
+			</div>
+		</div>